CHATBOT-PY/requirements.txt
langchain>=0.1.5
langchain-community>=0.0.17
langchain-openai>=0.0.5
chromadb>=0.4.22
langchain-chroma>=0.0.6
# pysqlite3-binary is commented out as it may not be available on all platforms
# pysqlite3-binary>=0.5.0
openai>=1.10.0
python-dotenv>=1.0.0
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
PyPDF2>=3.0.1
python-multipart>=0.0.6
tiktoken>=0.5.2
streamlit>=1.31.0
feedparser>=6.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
crawl4ai>=0.1.0
textdistance[extras]==4.5.0
python-Levenshtein>=0.21.1